loggers and refactor
