+++ src/main/java/uk/gov/hmcts/reform/prl/models/dto/ccd/CCDPaymentServiceRequestUpdate.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.prl.models.dto.payment;
+package uk.gov.hmcts.reform.prl.models.dto.ccd;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
