+++ src/main/java/uk/gov/hmcts/reform/prl/services/RequestUpdateCallbackService.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.prl.services;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
@@ -9,9 +10,9 @@ import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
+import uk.gov.hmcts.reform.prl.models.dto.ccd.CCDPayment;
+import uk.gov.hmcts.reform.prl.models.dto.ccd.CCDPaymentServiceRequestUpdate;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
-import uk.gov.hmcts.reform.prl.models.dto.payment.CCDPayment;
-import uk.gov.hmcts.reform.prl.models.dto.payment.CCDPaymentServiceRequestUpdate;
 import uk.gov.hmcts.reform.prl.models.dto.payment.ServiceRequestUpdateDto;
 
 import java.time.LocalDateTime;
@@ -19,6 +20,7 @@ import java.time.LocalDateTime;
 import static uk.gov.hmcts.reform.prl.enums.OrchestrationConstants.CASE_TYPE;
 import static uk.gov.hmcts.reform.prl.enums.OrchestrationConstants.JURISDICTION;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class RequestUpdateCallbackService {
@@ -32,7 +34,8 @@ public class RequestUpdateCallbackService {
 
     public void processCallback(ServiceRequestUpdateDto serviceRequestUpdateDto) {
 
-
+        log.info("Processing the callback for the caseId {} with status {}",serviceRequestUpdateDto.getCcdCaseNumber(),
+                 serviceRequestUpdateDto.getServiceRequestStatus());
         if (serviceRequestUpdateDto.getServiceRequestStatus().equalsIgnoreCase(PAID)) {
             CaseData caseData = setCaseData(serviceRequestUpdateDto);
